Initialize React application
Add basic output from create-react-app

Add a gitignore with standard things to ignore

Include yarn.lock since yarn will be the package manager.
Yarn was chosen for its improved performance over npm
as it is known to run more quickly, has a robust API,
is widely adopted (used by the React core project team),
and guarantees that running install on a project with a lockfile
will always produce the same tree of packages, unlike npm.
Yarn also plays nice with loom, in case this project goes the route
of a multi-package mono-repository. With yarn 2 around the corner,
it will also be easy to migrate, should that decision arise.