Add USWDS
Use trussworks/react-uswds because it integrates nicely and is well supported/maintained.

It became necessary to downgrade the Node.js version from 14 to 12
because the trussworks project only supports 12.
- Change Dockerfile node version to 12
- Create an .nvmrc file for folks who use nvm. This helps to ensure that
  a contributor's local node environment matches the Docker runtime environment.

Add a reference to USWDS in the README

Add README instructions to `build`, `run`, and `push` the Docker image
both for local development and to serve the static build.

Add a nice little UI to welcome people and to demonstrate that USWDS is integrated.

Remove unnecessary default create-react-app output files and code

Change the app title to match the project