Add ESLint and Prettier
Configure ESLint to work with Prettier
and follow the Airbnb style guide

Fix files to comply with ESLint/Prettier rules

Add .eslintignore and .prettierignore to omit files and directories
that should not be changed or checked

Document lint tools in the README

Add a yarn run v1.22.4
$ eslint src/ && echo 'Lint complete.'
Lint complete.
Done in 1.04s. command that can be used locally and by CI

As a note: Paulo Ramos' script was used to setup ESLint, Prettier,
and Airbnb config to play nice with each other. This is typically
an arduous process and the script made it simple.
For reference: https://github.com/paulolramos/eslint-prettier-airbnb-react