Implement Authentication and Authorization
Add Private routing which redirects unauthenticated users
back to homepage at '/' (Welcome component).

If API call to verify authentication is loading,
render an alert to notify the user it's loading.

If unauthenticated user tries to navigate to
a private route, redirect to unauthenticated home
Welcome component and render an Alert that user
needs to login first.

When the window first loads, make a call to retrieve
authentication info -- when authenticated, the user will
already have an http_only cookie set in the browser,
so the result of this call will identify the user as
already authenticated / or not yet authenticated
based on whether there is a valid cookie or not.

Add a Dashboard for authenticated users.

On the Welcome page, when user click 'Sign In',
user is redirected to API's Login.gov login flow.

Alerts are hooked up to the redux store so they can render
across the site regardless of the current page. This helps
preserve messages after redirects.

Additionally, the redux logger middleware order has changed
so logger comes last. If it is not last, it does not log
actions but instead logs promises and thunks:
https://github.com/LogRocket/redux-logger/issues/302#issuecomment-416767088