Add a Login Callback component
The API will redirect back to the web app
after the login flow has completed, delivering with it
an http_only cookie.  Upon landing on the callback page,
the web app makes one final call to the API to verify
that the web app indeed has the valid cookie. While this
happens, the login component renders a loading message.

If user is not authenticated, and not in the process
of doing so, then redirect from the callback page
to the Welcome homepage at '/'

If user is signed in, then redirect from callback page
to the Dashboard at '/dashboard'

Render loading message on any private route
if sign-in is currently in process.

Arrange Routes so public reoutes come first
because it's easier to reason about, and specify exact
so the order doesn't matter (no risk of unintended
fallthrough routing)

Redirect Welcome to Dashboard for signed in users
so they don't see a log-in button or any content
intended for public users.

Update redux actions and reducers so all data
is passed in a payload object, keeping the
action/reducer api consistent, predictable, and scalable

Add user email to the login response object, and render that
on the authenticated Dashboard.