Remove default 401 response
The frontend will call /auth_check on every page load, including
when the user is not signed in. The endpoint should not error
with a 401 but rather should return 200 and inform the client
whether or not the user is authenticated. The client is always
authorized to make this request.

Handle the case of anonymous users and those who are authenticated.